+++ qabelbox/src/androidTest/java/de/qabel/qabelbox/ui/SplashUITest.java
@@ -13,8 +13,8 @@ import org.junit.Test;
 
 import java.io.IOException;
 
-import de.qabel.qabelbox.activities.SplashActivity;
 import de.qabel.qabelbox.config.AppPreference;
+import de.qabel.qabelbox.startup.activities.SplashActivity;
 
 import static android.support.test.espresso.intent.matcher.ComponentNameMatchers.hasClassName;
 import static android.support.test.espresso.intent.matcher.IntentMatchers.hasComponent;