+++ qabelbox/src/main/java/de/qabel/qabelbox/reporter/CrashReporter.kt
@@ -1,0 +1,5 @@
+package de.qabel.qabelbox.reporter
+
+interface CrashReporter {
+    fun installCrashReporter()
+}